Flight controller is running on UDOO now!!

+++ src/libraries/Complementary_Filter/Comp_Filter.cpp
@@ -17,9 +17,9 @@ void Complementary_Filter::Calculate (float gyro_data, float accel_angle)
 {
     current_time = millis ();
      
-    dt = (current_time - last_time) / 1000.0;
+    dt = (float)((current_time - last_time) / 1000.0);
     
-    angle_radians = ((gC * (angle_radians + (gyro_data * (float)dt))) + (aC * accel_angle));
+    angle_radians = ((gC * (angle_radians + (gyro_data * dt))) + (aC * accel_angle));
     angle = angle_radians * 64.0;
 
     last_time = current_time;
